requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-binance==1.0.19
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
requests==2.32.3